+++ src/main/java/ru/vyarus/dropwizard/guice/module/installer/FeatureInstaller.java
@@ -1,15 +1,20 @@
 package ru.vyarus.dropwizard.guice.module.installer;
 
-import io.dropwizard.setup.Environment;
-
 /**
  * Installer serve two purposes: find extension on classpath and properly install it
  * (in dropwizard or somewhere else). Each installer should work with single feature.
  * <p>Installers are not guice beans: they are instantiated during guice context start and used to register
  * additional beans in guice context.</p>
- * <p>If installer implements {@code CustomBindingInstaller} it's bind method called during injector creation.
- * Note that each extension is always binded by type.</p>
- * <p>After context start installer is called to properly install class (register health check, resource, etc.).</p>
+ * <p>Installer may choose from three possible types of installation:
+ * <ul>
+ * <li>{@link ru.vyarus.dropwizard.guice.module.installer.install.BindingInstaller} to apply specific
+ * guice bindings (called in process of injector creation, whereas other installer are called after)</li>
+ * <li>{@link ru.vyarus.dropwizard.guice.module.installer.install.TypeInstaller} to register class
+ * in environment or somewhere else</li>
+ * <li>{@link ru.vyarus.dropwizard.guice.module.installer.install.InstanceInstaller} to register instance
+ * in environment or somewhere else</li>
+ * </ul>
+ * Even if installer doesn't implement any of these types, extension will be still registered in guice.</p>
  *
  * @param <T> expected extension type (or Object when no super type (e.g. for annotated beans))
  */
@@ -24,13 +29,4 @@ public interface FeatureInstaller<T> {
      * @return true if extension recognized, false otherwise
      */
     boolean matches(Class<?> type);
-
-    /**
-     * Every found extension (during classpath scan) is instantiated with guice and pass here for
-     * further registration.
-     *
-     * @param environment environment object
-     * @param instance    extension instance
-     */
-    void install(Environment environment, T instance);
 }