+++ src/main/java/ru/vyarus/dropwizard/guice/module/installer/feature/JerseyProviderInstaller.java
@@ -2,6 +2,7 @@ package ru.vyarus.dropwizard.guice.module.installer.feature;
 
 import io.dropwizard.setup.Environment;
 import ru.vyarus.dropwizard.guice.module.installer.FeatureInstaller;
+import ru.vyarus.dropwizard.guice.module.installer.install.InstanceInstaller;
 import ru.vyarus.dropwizard.guice.module.installer.util.FeatureUtils;
 
 import javax.ws.rs.ext.Provider;
@@ -10,7 +11,7 @@ import javax.ws.rs.ext.Provider;
  * Jersey provider installer.
  * Looks for classes annotated with {@code @javax.ws.rs.ext.Provider} and register in environment.
  */
-public class JerseyProviderInstaller implements FeatureInstaller<Object> {
+public class JerseyProviderInstaller implements FeatureInstaller<Object>, InstanceInstaller<Object> {
 
     @Override
     public boolean matches(final Class<?> type) {
@@ -19,6 +20,7 @@ public class JerseyProviderInstaller implements FeatureInstaller<Object> {
 
     @Override
     public void install(final Environment environment, final Object instance) {
+        // register by instance to force singleton: prototype providers not allowed by jersey
         environment.jersey().register(instance);
     }
 }