+++ src/main/java/ru/vyarus/dropwizard/guice/module/installer/feature/LifeCycleInstaller.java
@@ -3,13 +3,14 @@ package ru.vyarus.dropwizard.guice.module.installer.feature;
 import io.dropwizard.setup.Environment;
 import org.eclipse.jetty.util.component.LifeCycle;
 import ru.vyarus.dropwizard.guice.module.installer.FeatureInstaller;
+import ru.vyarus.dropwizard.guice.module.installer.install.InstanceInstaller;
 import ru.vyarus.dropwizard.guice.module.installer.util.FeatureUtils;
 
 /**
  * Lifecycle objects installer.
  * Looks for classes implementing {@code org.eclipse.jetty.util.component.LifeCycle} and register them in environment.
  */
-public class LifeCycleInstaller implements FeatureInstaller<LifeCycle> {
+public class LifeCycleInstaller implements FeatureInstaller<LifeCycle>, InstanceInstaller<LifeCycle> {
 
     @Override
     public boolean matches(final Class<?> type) {