+++ src/main/java/ru/vyarus/dropwizard/guice/module/installer/feature/ManagedInstaller.java
@@ -3,13 +3,14 @@ package ru.vyarus.dropwizard.guice.module.installer.feature;
 import io.dropwizard.lifecycle.Managed;
 import io.dropwizard.setup.Environment;
 import ru.vyarus.dropwizard.guice.module.installer.FeatureInstaller;
+import ru.vyarus.dropwizard.guice.module.installer.install.InstanceInstaller;
 import ru.vyarus.dropwizard.guice.module.installer.util.FeatureUtils;
 
 /**
  * Managed objects installer.
  * Looks for classes implementing {@code io.dropwizard.lifecycle.Managed} and register them in environment.
  */
-public class ManagedInstaller implements FeatureInstaller<Managed> {
+public class ManagedInstaller implements FeatureInstaller<Managed>, InstanceInstaller<Managed> {
     @Override
     public boolean matches(final Class<?> type) {
         return FeatureUtils.is(type, Managed.class);