+++ src/main/java/ru/vyarus/dropwizard/guice/module/installer/feature/eager/EagerInstaller.java
@@ -3,6 +3,7 @@ package ru.vyarus.dropwizard.guice.module.installer.feature.eager;
 
 import io.dropwizard.setup.Environment;
 import ru.vyarus.dropwizard.guice.module.installer.FeatureInstaller;
+import ru.vyarus.dropwizard.guice.module.installer.install.InstanceInstaller;
 import ru.vyarus.dropwizard.guice.module.installer.util.FeatureUtils;
 
 /**
@@ -12,7 +13,7 @@ import ru.vyarus.dropwizard.guice.module.installer.util.FeatureUtils;
  * automatically. Moreover, even in DEVELOPMENT stage instance will be requested, which makes class suitable
  * for initialization logic.
  */
-public class EagerInstaller implements FeatureInstaller<Object> {
+public class EagerInstaller implements FeatureInstaller<Object>, InstanceInstaller<Object> {
     @Override
     public boolean matches(final Class<?> type) {
         return FeatureUtils.hasAnnotation(type, Eager.class);
@@ -20,6 +21,6 @@ public class EagerInstaller implements FeatureInstaller<Object> {
 
     @Override
     public void install(final Environment environment, final Object instance) {
-        // do nothing - instance already registered in guice (and jsr250 annotations may be used)
+        // instance installer required only to force bean instance creation (and so jsr250 annotations could act)
     }
 }