+++ src/test/groovy/ru/vyarus/dropwizard/guice/support/feature/CustomFeature.groovy
@@ -1,8 +1,11 @@
 package ru.vyarus.dropwizard.guice.support.feature
 
 /**
+ * Custom feature to build custom installer around.
+ *
  * @author Vyacheslav Rusakov 
  * @since 04.09.2014
+ * @see ru.vyarus.dropwizard.guice.support.installer.CustomInstaller
  */
 class CustomFeature {
 }