+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 * Add dependency on guice-multibindings
-* Add custom bindings definition in installer through BindingInstaller marker interface
+* Installers may choose now from three types of installation (binding, type or instance) or combine them.
 * Add PluginInstaller: shortcut for multibindings mechanism
+* Updated guice (4.0.beta4 -> 4.0.beta5)
 
 ### 0.9.0 (2014-09-05)